Make android player file default for now
